Refactor to simplify UITableViewCell creation.
Having a method + class (+ static Identifier in the class) is too much when
the method is boilerplate (one-line wrapper over GetCell<T>) and the
identifier can be inferred from the Section Title.

Drop the requirement to have a method, and use the Section Title as the
UITableViewCell identifier.

This way, we only need to add a new entry to Sections and a new UIView
subclass to provide additional samples.