Undo iPhoneOS 3.2 requirement.
Instead of requiring iPhoneOS 3.2, do a runtime check to see if the CoreText
framework exists.  If it does, show the demo; otherwise, hide it.

Note: at present, the only way to actually view the CoreText demo is to alter
the project file (e.g. for Debug|iPhoneSimulator) and set the SDK version to
3.2; there is no mechanism (at present) to build for 3.0 and run under 3.2.