Implement "proper" support for distinct iPad & iPhone UIs.
For "cleanliness" (and sanity), place the iPhone UI files into the ui-iPhone
project, and add a ui-iPad project (for the iPad UI).  Add a 'core' project
for types which should be shared between the distinct UIs.

Unfortunately, MonoDevelop doesn't support placing .xib files into separate
projects, so we use a post-build step to generate the .nib files and place
them into the appropriate output directory.

To run the (non-existent) iPad UI, change the monocatalog project to target
the iPhoneOS 3.2 SDK.  This will cause monocatalog to be loaded within the
iPad simulator and use the iPad .nib files for the UI.

By default we start the iPhone UI.