Get something actually loading on the iPad.
More or less, anyway.

Problem 1: the File's Owner in the .xib hadn't set the delegate property to
SplitViewAppDelegate.  Result: only a whitescreen on load, and nothing
happened.  Oops.

Problem 2: Info.plist didn't have a UISupportedInterfaceOrientations key.
Result: when rotating the iPad (which is necessary to view the menu), nothing
happened -- meaning it was ~useless.  List all possible positions as valid.

This isn't perfect: if you start in a portrait position you don't see anything
(so switch to landscape), and the CoreText demo is shown in the left-most menu
area, not in the "detail view" on the right.  Fixes to come...