Attempt to get a real iPad UI working...
...and failing.

So, the good news: now, if you view in landscape you'll actually get to see
the UISplitViewController in all its glory, with a single entry.  This is the
good news.

The bad news is that if you select that single entry, the entire screen is
replaced, and the UISplitViewController ceases to work.  Oops.

Another aborted attempt (uncommitted; described for background) was, instead
of leaving the "DetailViewController" blank in MainWindow_Pad.xib, to instead
provide a UINavigationController and then call
NavigationController.PushViewController() to add the actual
CoreTextController_Pad controller.

This worked, and worked nicely, except for one "minor" detail:

Because we were calling PushViewController(), the NavigationController's
"stack" always had > 1 element.  Consequently, in landscape mode we always had
a "back" button...which was pointless, as there was nowhere to go back *to*
(as the left "pane" of the UISplitViewController served the purpose of what
used to be the first NavigationController page...).

Attempting to workaround this by using
ParentViewController.PresentModalViewController() failed horribly, causing it
to...replace th UISplitViewController.

Hm...notice a pattern?