Get the iPad UI into a reasonable state.
Now, as expected (expectations based on the behavior of the Xcode
SplitView-based application), when we start in portrait orientation we show a
Demos button in the top-left corner of the screen which allows selecting the
Demo to view.  Rotate into a landscape orientation and this button disappears
while a menu fills the left part of the screen.

In short, things Just Work.